Support for conversion errors and default attribute weights
The service now keeps reading data despite conversion errors, but ignores rows where errors occurred.  This behaviour can be disabled by setting config.ignoreConversionErrors = true.

The default value of distWeight is no longer 1, but 1 / (variance of the attribute).  For the purpose of calculating the variance, the highest and lowest few values are ignored; this can be controlled via config.distWeightOutliers (default: 0.05, meaning that the highest and lowest 2.5% of the values are ignored).