State history added
The modelling service now outputs two arrays which show at what points in time the system changes from one initial state to another.  These arrays increase the size of the JSON response considerably; to avoid increasing it still further, equivalent arrays for higher scales are not included, since the caller can easily translate the identifiers of initial states into the identifiers of their ancestors at higher scales.

Furthermore, the config object of the input JSON now supports flags that allow the caller to disable the generation of decision trees, histograms, and state history.  This might be useful to save time when testing.