Classification support
The service now supports the /classifySamples function, which takes as input a model and a set of datapoints, and outputs, for each of these datapoints, the number of the initial state whose centroid was the closest to that datapoint.

The JSON represetation of the model returned by the /buildModel function now contains additional structures required by /classifySamples.  Thus /classifySamples will not work with models produced by earlier versions of the service.

Each state object in the JSON output of /buildModel now contains a boolean field named "sameAsParent" which indicates whether the state is identical to its parent.  If yes, the state will not include the following fields: histograms, centroid, decisionTree, suggestedLabel.  The caller should copy them from the parent state instead.  This significantly reduces the size of the JSON representation of the model.

A new version of sampleCall/testCall.py is provided which now also shows how to use the new /classifySamples function.